# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/serviceprovider": {
            "get": {
                "description": "This API List all Service Providers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Provider"
                ],
                "summary": "This API List all Service Providers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.FPServiceProvider"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/serviceprovider/add": {
            "post": {
                "description": "This API Creates a new Service Provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Provider"
                ],
                "summary": "This API Creates a new Service Provider",
                "parameters": [
                    {
                        "description": "New Service Provider",
                        "name": "serviceprovider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FPServiceProvider"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FPServiceProvider"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/serviceprovider/delete/{serviceproviderId}": {
            "delete": {
                "description": "This API Delete Service Provider by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Provider"
                ],
                "summary": "This API Delete Service Provider by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of ServiceProvider",
                        "name": "serviceProviderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/serviceprovider/update": {
            "post": {
                "description": "This API Updates a Service Provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Provider"
                ],
                "summary": "This API Updates a Service Provider",
                "parameters": [
                    {
                        "description": "Existing Service Provider",
                        "name": "serviceprovider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FPServiceProvider"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FPServiceProvider"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/serviceprovider/{serviceproviderId}": {
            "get": {
                "description": "This API Finds Service Provider by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Provider"
                ],
                "summary": "This API Finds Service Provider by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of Service Provider",
                        "name": "serviceProviderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FPPremium"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.FPAPIResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "models.FPPremium": {
            "type": "object",
            "properties": {
                "fppremiumDescription": {
                    "type": "string"
                },
                "fppremiumID": {
                    "type": "string"
                },
                "fppremiumName": {
                    "type": "string"
                },
                "fppremiumServiceID": {
                    "type": "string"
                },
                "fppremiumStatus": {
                    "type": "string"
                }
            }
        },
        "models.FPServiceProvider": {
            "type": "object",
            "properties": {
                "fpserviceProviderAddress": {
                    "type": "string"
                },
                "fpserviceProviderContactNumber": {
                    "type": "string"
                },
                "fpserviceProviderEmail": {
                    "type": "string"
                },
                "fpserviceProviderFspNumber": {
                    "type": "string"
                },
                "fpserviceProviderID": {
                    "type": "string"
                },
                "fpserviceProviderName": {
                    "type": "string"
                },
                "fpserviceProviderStatus": {
                    "type": "string"
                }
            }
        },
...
    }
}